+++ upkie/cpp/observers/tests/observe_servos_test.cpp
@@ -5,18 +5,20 @@
 
 #include <palimpsest/Dictionary.h>
 
+#include <limits>
 #include <map>
 #include <string>
 #include <vector>
 
 #include "gtest/gtest.h"
 #include "upkie/cpp/actuation/moteus/ServoReply.h"
+#include "upkie/cpp/exceptions/ServoError.h"
 
 namespace upkie::cpp::observers {
 
-TEST(Servo, ReadTorques) {
-  palimpsest::Dictionary observation;
+using upkie::cpp::exceptions::ServoError;
 
+TEST(Servo, ReadTorques) {
   std::map<int, std::string> servo_joint_map = {{0, "foo"}, {1, "bar"}};
   std::vector<actuation::moteus::ServoReply> servo_replies;
   servo_replies.push_back({1, {}});           // bar first
@@ -24,8 +26,8 @@ TEST(Servo, ReadTorques) {
   servo_replies.push_back({0, {}});           // foo next
   servo_replies.back().result.torque = 10.;   // [N m]
 
+  palimpsest::Dictionary observation;
   observe_servos(observation, servo_joint_map, servo_replies);
-
   ASSERT_TRUE(observation.has("servo"));
   ASSERT_TRUE(observation("servo").has("foo"));
   ASSERT_TRUE(observation("servo").has("bar"));