+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update downloaded simulation spine in cache if outdated
 - actuation: Added link inertia randomization to the Bullet interface (thanks to @Tordjx)
 - actuation: Check maximum torques before sending commands
+- cpp: Add `ServoError` exception
 - model: Static model for joints with position, velocity and torque limits
 - pixi: Task to generate the documentation by `pixi -e docs make`